backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.4.4
packaging==23.1
pyspark==3.3.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
